+++ CHANGELOG.md
@@ -34,8 +34,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `console:print_err/1` to write to standard error
 - Added `erlang:term_to_binary/2`, `erlang:is_builtin/3` and `erlang:bitstring_to_list/1`
 - Added `lists:mapfoldr/3`
-- Added `maps:take/2`, `maps:update_with/3`, `maps:update_with/4`, `maps:with/2` and
-  `maps:without/2` to the estdlib `maps` module
+- Added `maps:take/2`, `maps:update_with/3`, `maps:update_with/4`, `maps:with/2`,
+  `maps:without/2`, `maps:filtermap/2`, `maps:intersect/2`, `maps:intersect_with/3`,
+  `maps:groups_from_list/2` and `maps:groups_from_list/3` to the estdlib `maps` module
 
 ### Changed
 - Updated network type db() to dbm() to reflect the actual representation of the type